requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
numpy>=1.24.0